mixer: add initial support for ALSA mixer
The ALSA mixer interface implements a simple linear volume control for
hardware mixers.

Note that a linear volume control is not ideal for human audio
experience as our ears respond to sound on an exponential curve.

The mixer also works fine with the ALSA softvol plugin which offers an
logarithmic attenuation curve.

TODO: Use logarithmic curve for hardware mixers.